[:micro:metadata] Move GatewayHeadersBuilder from [:core]
