+++ http/server/src/main/java/io/zero88/qwe/http/server/handler/DynamicContextDispatcher.java
@@ -7,7 +7,6 @@ import io.vertx.core.http.HttpMethod;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.web.RoutingContext;
 import io.zero88.qwe.dto.ErrorMessage;
-import io.zero88.qwe.dto.msg.GatewayHeadersBuilder;
 import io.zero88.qwe.dto.msg.RequestData;
 import io.zero88.qwe.dto.msg.RequestFilter;
 import io.zero88.qwe.dto.msg.ResponseData;
@@ -16,6 +15,7 @@ import io.zero88.qwe.http.HttpStatusMapping;
 import io.zero88.qwe.http.HttpUtils;
 import io.zero88.qwe.http.server.converter.RequestDataConverter;
 import io.zero88.qwe.http.server.rest.api.DynamicRestApi;
+import io.zero88.qwe.micro.GatewayHeadersBuilder;
 import io.zero88.qwe.micro.ServiceDiscoveryApi;
 import io.zero88.qwe.micro.filter.ByPredicateFactory;
 import io.zero88.qwe.micro.filter.RecordPredicateFactory;
