+++ micro/discovery/src/test/java/io/zero88/qwe/rpc/GatewayServiceInvokerTest.java
@@ -11,12 +11,12 @@ import io.vertx.ext.unit.TestContext;
 import io.vertx.junit5.VertxTestContext;
 import io.zero88.qwe.JsonHelper.Junit4;
 import io.zero88.qwe.TestHelper;
-import io.zero88.qwe.dto.msg.GatewayHeadersBuilder;
 import io.zero88.qwe.dto.msg.RequestData;
 import io.zero88.qwe.event.EventAction;
 import io.zero88.qwe.exceptions.ErrorCode;
 import io.zero88.qwe.exceptions.QWEException;
 import io.zero88.qwe.micro.BaseDiscoveryPluginTest;
+import io.zero88.qwe.micro.GatewayHeadersBuilder;
 import io.zero88.qwe.micro.mock.MockServiceListener;
 import io.zero88.qwe.rpc.mock.MockServiceInvoker;
