+++ micro/metadata/src/main/java/io/zero88/qwe/micro/filter/EventMessageServicePathPredicate.java
@@ -3,9 +3,9 @@ package io.zero88.qwe.micro.filter;
 import io.vertx.core.http.HttpMethod;
 import io.vertx.core.json.JsonObject;
 import io.vertx.servicediscovery.Record;
-import io.zero88.qwe.dto.msg.GatewayHeadersBuilder;
 import io.zero88.qwe.event.EventAction;
 import io.zero88.qwe.http.EventMethodDefinition;
+import io.zero88.qwe.micro.GatewayHeadersBuilder;
 import io.zero88.qwe.micro.servicetype.EventMessageHttpService;
 import io.zero88.qwe.utils.JsonUtils;
